Add reducer
